feature: completes top up wallet command.
Adds msg property to Checkout page metadata
