+++ bot/src/commands/wallet/balance/types.ts
@@ -1,0 +1,7 @@
+export type BalanceReplyProps = {
+  currencyCode: string;
+  currentBalance: number;
+  onHoldBalance: number;
+  reserveBalance: number;
+  receivedBalance: number;
+};
